Assignment 2/requirements.txt
tensorflow
pytest
numpy
PyYAML
matplotlib
tensorflow-metal
tqdm
scikit-learn